% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractCmDataObjects.R
\name{extractCohortMethodData}
\alias{extractCohortMethodData}
\title{Adopted from runCohortMethod in this package to more cleanly extract data objects we need.}
\usage{
extractCohortMethodData(
  outputFolder,
  indicationId = "cde",
  databaseId,
  maxCores = 4,
  includeNegativeControls = FALSE,
  runSections
)
}
\description{
Adopted from runCohortMethod in this package to more cleanly extract data objects we need.
}
